Rename registries folder to entities
