+++ src/app/core/entities/entities.ts
