+++ src/app/core/entities/types.ts
