+++ src/app/core/heroes/types.ts
@@ -6,7 +6,7 @@ import { Item, ItemBaseModel } from '../items';
 import { InventoryItems } from '../items/inventory';
 import { Modifiers, ModsRef, ModsRefsGroup, Specialties, filterSpecialties } from '../modifiers';
 import { Player } from '../players';
-import { Entity, HeroId } from '../registries';
+import { Entity, HeroId } from '../entities';
 import { ResourcesModel } from '../resources';
 import { Spell, SpellBaseType } from '../spells';
 import { DescriptionElement } from '../ui';
