+++ src/app/core/structures/common/lock-location.ts
@@ -1,4 +1,4 @@
-import { StructId } from '../../registries';
+import { StructId } from '../../entities';
 import { Resources, joinedResourcesText } from '../../resources';
 import { StructureGeneratorModel, StructureType, StuctureControl } from '../types';
 import { createStructure } from '../utils';
