+++ src/app/core/structures/common/resource-mine.ts
@@ -1,4 +1,4 @@
-import { StructId } from '../../registries';
+import { StructId } from '../../entities';
 import { ResourceType, Resources, resourceNames } from "../../resources";
 import { StructureGeneratorModel, StructureType, StuctureControl } from "../types";
 import { createStructure } from '../utils';
