+++ src/app/core/towns/types.ts
@@ -1,7 +1,7 @@
 import { NewWeekStarted } from '../events';
 import { GameObject } from '../game-objects';
 import { ItemBaseModel } from '../items';
-import { Entity, TownId } from '../registries';
+import { Entity, TownId } from '../entities';
 import { ActivityTypes, Building, BuildingDescription, HiringActivity } from './buildings';
 
 export interface TownBase<T extends string> extends Entity {
