+++ src/app/core/unit-types/types.ts
@@ -7,7 +7,7 @@ import { Hero } from '../heroes';
 import { ModsRef, ModsRefsGroup, Specialties } from '../modifiers';
 import { Modifiers } from '../modifiers/modifiers';
 import type { Player } from '../players';
-import { Entity, EntityId, UnitId } from '../registries';
+import { Entity, EntityId, UnitId } from '../entities';
 import { ResourcesModel } from '../resources';
 import { Spell, SpellBaseType } from '../spells';
 import { DescriptionElement } from '../ui';
